ライン結合のバグを修正、デバックしやすいようにソースファイルを分割
