+++ src/HWCAM.py
@@ -26,976 +26,27 @@ Created on Sat Jun 27 13:29:53 2020
 # os
 # datetime
 
+# 外部ライブラリ
 import tkinter as tk
-import tkinter.ttk as ttk
-import ezdxf as ez
 import numpy as np
 from matplotlib import pyplot as plt
 from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg, NavigationToolbar2Tk
 from mpl_toolkits.mplot3d import Axes3D
-from scipy import interpolate as intp
 from matplotlib.figure import Figure
-from scipy.integrate import quad
-import os.path as op
 import datetime
 import os
-import sys
 import traceback
 
+# 内部ライブラリ
+from cam_generic_lib import *
+from line_object import *
+from dxf_file import *
+from messeage_window import *
+from cam_global import *
 
-#======================================================================================================================================
-#            グローバル変数
-#======================================================================================================================================
-VERSION = "3.0"           #バージョン
-Z_ERROR = 5.0             #2020.10.13　ver2.1 新規追加  単位：mm Z面の許容誤差 
-LINE_MARGE_NORM_MN = 1    #単位:mm ラインを結合時にラインを結合してよいかを判断するためのライン端点間距離
-N_FILLET_INTERPORATE = 10 #オフセットした線間を補完するフィレットの座標点数
-
-   
 #======================================================================================================================================
 #            クラスの実装
 #======================================================================================================================================
-
-
-#######################################################################################################################################
-###############    line_objectクラス   ここから　　　#########################################################################################
-
-#【説明】
-#　ライン，スプラインを格納するクラスである．
-#
-#【親クラス】
-#　なし
-#
-#【メンバ変数】
-#   変数名          型              説明
-#   x_raw           list            x座標点列の初期値　変更されない
-#   y_raw           list            y座標点列の初期値　変更されない
-#   st              list            x_raw, y_rawの 始点 
-#   ed              list            x_raw, y_rawの 終点 
-#   N               int             座標点列の個数
-#   num             int             ラインの番号
-#   offset_dist     float           オフセット距離
-#   offset_dir      char            オフセット方向["O"/ "I"]
-#   cut_dir         char            x, y配列の方向["F"/ "R"]（x_raw，y_rawに対して順方向が"F"，逆方向が"R"）
-#   x               list            オフセット，配列方向を適用したあとの x座標点列
-#   y               list            オフセット，配列方向を適用したあとの y座標点列
-#
-#【実装メソッド】
-#   __init__(list x_points, list y_points, int num)
-#   【引数】 x_points, y_points, num
-#   【戻り値】　なし
-#   【機能】 初期化処理を実行する．x座標点列x_points, y座標点列y_points　を，x_raw，y_rawおよびx, yに代入する．ライン名をnumに設定する．
-#
-#   update()
-#   【引数】　なし
-#   【戻り値】　なし
-#   【機能】 cut_dir, offset_dir, offset_distの現在値をもとに，x, yの配列を再生成する
-#
-#   set_offset_dir(char offset_dir)
-#   【引数】 offset_dir
-#   【戻り値】　なし
-#   【機能】　メンバ変数であるoffset_dirを入力値に設定する．offset_dirが"O", "I"以外は処理しない．
-#
-#   set_offset_dist(float offset_dist)
-#   【引数】 offset_dist
-#   【戻り値】 なし
-#   【機能】 メンバ変数であるoffset_distを入力値に設定する．
-#
-#   set_cut_dir(char cut_dir)
-#   【引数】 cut_dir 
-#   【戻り値】 なし
-#   【機能】 メンバ変数であるcut_dirtを入力値に設定する．cut_dirが"F", "R"以外は処理しない．
-#
-#   set_num(int num)
-#   【引数】　num 
-#   【戻り値】　なし
-#   【機能】 メンバ変数であるnumを入力値に設定する．
-#
-#   toggle_cut_dir()
-#   【引数】 なし
-#   【戻り値】　なし
-#   【機能】　カット方向を逆転させる
-#
-#   toggle_offset_dir()
-#   【引数】　なし
-#   【戻り値】　なし
-#   【機能】　オフセット方向を逆転させる
-#
-#   calc_length_array(str mode = "offset")
-#   【引数】 mode（オプション）
-#   【戻り値】　list length_array
-#   【機能】　x, yのスプラインに沿った線長を，配列として返す．mode = "raw"の場合，x_raw，y_rawのスプラインに沿った線長を配列として返す．
-#
-#   get_length(str mode = "offset")
-#   【引数】 mode（オプション）
-#   【戻り値】 float length
-#   【機能】 x, yのスプラインに沿った線長を返す．mode = "raw"の場合，x_raw，y_rawのスプラインに沿った線長を返す．
-
-
-class line_object:
-    def __init__(self, x_points, y_points, num):
-        self.x_dxf = np.array(x_points)
-        self.y_dxf = np.array(y_points)
-        self.x_raw = np.array(x_points)
-        self.y_raw = np.array(y_points)
-        self.st = np.array([x_points[0], y_points[0]])
-        self.ed = np.array([x_points[-1], y_points[-1]])
-        self.N = max(len(x_points), len(y_points))
-        self.num = num
-        self.offset_dist = 0
-        self.offset_dir = "O"
-        self.cut_dir = "F"
-        self.cutspeed_work = 100
-        self.cutspeed_mech = 100
-        self.x = x_points
-        self.y = y_points
-        self.interp_mode = "cubic" #ver.2.2追加 "cubic":3d-spline, "linear":1d-line, ポリラインの指定用
-        self.offset_ox = 0
-        self.offset_oy = 0
-
-        if len(x_points)<2:
-            self.line_type = "point"
-        if len(x_points)==2:
-            self.line_type = "line"
-        if len(x_points)>2:
-            self.line_type = "spline"
-       
-        
-    def reset_point(self, x_points, y_points, offset_ox, offset_oy):
-        self.x_dxf = np.array(x_points)
-        self.y_dxf = np.array(y_points)
-        self.offset_ox = offset_ox
-        self.offset_oy = offset_oy
-        self.x_raw = np.array(x_points) + offset_ox
-        self.y_raw = np.array(y_points) + offset_oy
-        self.st = np.array([x_points[0], y_points[0]]) + offset_ox
-        self.ed = np.array([x_points[-1], y_points[-1]]) + offset_oy
-        self.N = max(len(x_points), len(y_points))
-        self.x = np.array(x_points) + offset_ox
-        self.y = np.array(y_points) + offset_oy
-        self.interp_mode = "cubic" 
-
-        if len(x_points)<2:
-            self.line_type = "point"
-        if len(x_points)==2:
-            self.line_type = "line"
-        if len(x_points)>2:
-            self.line_type = "spline"        
-    
-    
-    def update(self):
-        if self.cut_dir == "R":
-            temp_x = self.x_raw[::-1]
-            temp_y = self.y_raw[::-1]
-        else:
-            temp_x = self.x_raw
-            temp_y = self.y_raw
-        
-        if self.offset_dir == "I":
-            self.x, self.y = offset_line(temp_x, temp_y, self.offset_dist, self.cut_dir, self.interp_mode) 
-            
-        else:
-            self.x, self.y = offset_line(temp_x, temp_y, -self.offset_dist, self.cut_dir, self.interp_mode)
-    
-    
-    def set_offset_dir(self, offset_dir):
-        if offset_dir == 'O' or offset_dir == 'I':
-            self.offset_dir = offset_dir
-            self.update()
-        
-        
-    def set_offset_dist(self, offset_dist):
-        try:             
-            self.offset_dist = float(offset_dist)
-            self.update()
-        except:
-            pass
-    
-    
-    def set_cut_dir(self, cut_dir):
-        if cut_dir == 'F' or cut_dir == 'R':
-            self.cut_dir = cut_dir
-            self.update()
-        
-            
-    def set_cutspeed(self, cutspeed_work, cutspeed_mech):
-        self.cutspeed_work = cutspeed_work
-        self.cutspeed_mech = cutspeed_mech
-        
-        
-    def set_num(self, num):
-        try:
-            self.num = int(num)
-        except:
-            pass
-    
-    
-    def toggle_cut_dir(self):
-        if self.cut_dir == "F":
-            self.cut_dir = "R"
-        else:
-            self.cut_dir = "F"
-    
-    
-    def toggle_offset_dir(self):
-        if self.offset_dir == "O":
-            self.offset_dir = "I"
-        else:
-            self.offset_dir = "O"
-
-        
-    def calc_length_array(self, mode = "offset"):
-        
-        length_array = [0]
-        
-        if mode == "raw":
-            x = self.x_raw
-            y = self.y_raw
-        
-        else:
-            x = self.x
-            y = self.y           
-            
-        if self.line_type == "point":
-            length_array = []
-        
-        if self.line_type == "line":
-            dl = np.sqrt((x[0]-x[1])**2 + (y[0]-y[1])**2)
-            length_array.append(dl)
-        
-        if self.line_type == "spline":
-            dl = 0
-            dt = 1
-            s_l = 0
-            t_p = np.linspace(0, dt * len(x), len(x))
-
-            fx_t = intp.CubicSpline(t_p, x)
-            fy_t = intp.CubicSpline(t_p, y)
-            
-            dfx_t = fx_t.derivative(1)
-            dfy_t = fy_t.derivative(1)
-            
-            fdl = lambda t: np.sqrt(dfx_t(t)**2 + dfy_t(t)**2)
-            
-            i = 0
-            while i < len(t_p) - 1:
-                dl = quad(fdl, t_p[i], t_p[i+1])
-                s_l += dl[0]
-                length_array.append(s_l)
-                i += 1
-            
-        return np.array(length_array)
-    
-    
-    def get_length(self, mode = "offset"):
-        
-        length = 0
-        
-        if mode == "raw":
-            x = self.x_raw
-            y = self.y_raw
-        
-        else:
-            x = self.x
-            y = self.y           
-            
-        if self.line_type == "point":
-            length = 0
-        
-        if self.line_type == "line":
-            length = np.sqrt((x[0]-x[1])**2 + (y[0]-y[1])**2)
-        
-        if self.line_type == "spline":
-            dl = 0
-            dt = 1
-            s_l = 0
-            t_p = np.linspace(0, dt * len(x), len(x))
-
-            fx_t = intp.CubicSpline(t_p, x)
-            fy_t = intp.CubicSpline(t_p, y)
-            
-            dfx_t = fx_t.derivative(1)
-            dfy_t = fy_t.derivative(1)
-            
-            fdl = lambda t: np.sqrt(dfx_t(t)**2 + dfy_t(t)**2)
-            
-            i = 0
-            while i < len(t_p) - 1:
-                dl = quad(fdl, t_p[i], t_p[i+1])
-                s_l += dl[0]
-                i += 1
-            length = s_l
-        
-        return length
-
-###############    line_objectクラス   ここまで　　　#########################################################################################
-#######################################################################################################################################
-
-
-
-#######################################################################################################################################
-###############    super_tableクラス   ここから　　　#########################################################################################
-        
-#【説明】
-#　ラインのテーブルを作成するクラスである．tk.treeのインスタンスを格納する．tk.treeを初期化する方法がtreeの再生成しかないため，生成に必要な情報をメンバ変数として保存し，
-#　reset()関数がコールされた際にtableを削除，再生成する．注意：tk.treeは親クラスではない．
-#
-#【親クラス】
-#　なし
-#
-#【メンバ変数】
-#   変数名          型              説明
-#   x_pos           int            テーブルの配置位置 X座標
-#   y_pos           int            テーブルの配置位置 Y座標
-#   root            tk.Frame       テーブル先のフレーム
-#   y_height        int            テーブル先の高さ（行）
-#   parent          int            テーブルの親子設定[0 / 1]（1：親，0：子）
-#   table           tk.tree        テーブルのインスタンス
-#
-#【実装メソッド】
-#   __init__(tk.Frame root, int y_height, int x, int y)
-#   【引数】 root, y_height, x, y
-#   【戻り値】　なし
-#   【機能】 クラスを初期化する．
-#
-#   set_parent(int parent):
-#   【引数】　parent
-#   【戻り値】　なし
-#   【機能】 parentを入力値に変更する
-#
-#   reset()
-#   【引数】 なし
-#   【戻り値】　なし
-#   【機能】　既存のtableインスタンスを破棄する．その後再生成する．テーブルのヘッダなどを設定する．
-
-    
-class super_table:
-    def __init__(self, root, y_height, x, y):
-        self.x_pos = x
-        self.y_pos = y
-        self.root = root
-        self.y_height = y_height
-        self.parent = 0
-        self.reset()
-    
-    
-    def set_parent(self, parent):
-        if parent == 1 or parent == 0:
-            self.parent = parent
-    
-    
-    def reset(self):
-        try:
-            self.table.destroy()
-        except:
-            pass
-        self.table = ttk.Treeview(self.root, height = self.y_height)
-        self.table.place(x = self.x_pos, y = self.y_pos)
-        self.table["column"] = (1,2,3,4,5,6)
-        self.table["show"] = "headings"
-        self.table.heading(1,text="番号")
-        self.table.heading(2,text="カット方向")
-        self.table.heading(3,text="オフセット方向")
-        self.table.heading(4,text="オフセット距離")
-        self.table.heading(5,text="タイプ")
-        self.table.heading(6,text="カット速度")
-        self.table.column(1, width=40)
-        self.table.column(2, width=60)
-        self.table.column(3, width=80)
-        self.table.column(4, width=80)
-        self.table.column(5, width=60)
-        self.table.column(6, width=70)
-    
-###############    super_tableクラス   ここまで　　　#########################################################################################
-#######################################################################################################################################
-
-
-
-#######################################################################################################################################
-###############   dxf_fileクラス   ここから　　    　#########################################################################################
-
-#【説明】
-#　dxfファイルに関連する，line_object，super_tableなどを格納する．main関数以外ではトップレベルモジュールである．
-#
-#【親クラス】
-#　なし
-#
-#【メンバ変数】
-#   変数名          型                   説明
-#   ax              Axes                グラフのプロットエリアのインスタンス
-#   canvas          FigureCanvasTkAgg   グラフを配置しているフレームのインスタンス
-#   table           super_table         x_raw, y_rawの 始点 
-#   x_table         super_table         x_raw, y_rawの 終点 
-#   name            str                 グラフに表示する名前"X-Y", "U-V"など
-#   line_list       list                line_objectをすべて格納するリスト
-#   line_num_list   list                テーブル番号とline_objectを対応付けるリスト
-#   selected_line   int                 テーブルにて選択されたラインの番号
-#   filename        str                 dxfファイルの名称
-#
-#【実装メソッド】
-#   __init__(Axes ax, FigureCanvasTkAgg canvas, super_table table, super_table x_table, str name)
-#   【引数】 x_points, y_points, num
-#   【戻り値】　なし
-#   【機能】 初期化処理を実行する．ax，canvas，table，x_table，nameを自身のメンバ変数に格納する，またline_list，line_num_listを空のリストとして作成する．
-#　　　　　　　　またselected_numを0として作成する．
-#
-#   load_file(str filename)
-#   【引数】　なし
-#   【戻り値】　なし
-#   【機能】 filenameをメンバ変数にかくのうし，他のメンバ変数を初期化する，新しいdxfファイルを読み込む．reload()，plot()をコールする，
-#　　　　　　　　テーブルの選択イベントと，selectedをバインドする．
-#
-#   reload()
-#   【引数】 なし
-#   【戻り値】　なし
-#   【機能】　filemameのdxfファイルを読み込み，lineとsplineの座標点列を取得する．座標点列からline_objectを生成し，line_listに追加する．
-#　　　　　　　　line_num_listに自身の番号を追加する．（初期値はline_object.numと同じ）　tabelにline_objectの情報を追加する．
-#
-#   table_reload()
-#   【引数】 なし
-#   【戻り値】 なし
-#   【機能】 tableに表示する情報を，line_objectのメンバ変数が持つ情報に更新する，
-#
-#   plot()
-#   【引数】 なし 
-#   【戻り値】 なし
-#   【機能】 グラフを更新する．グラフは，selected_line以外を透明度を低下させて描画する．グラフは，selected_lineの始点からカット方向ベクトルを描画する．
-#　　　　　　　　グラフは，offset_distが 0以外である場合，オフセット方向ベクトルを描画する．グラフは，item_num_listの値が0以上の場合にのみプロットする．
-#
-#   set_selected_line(int selected_line)
-#   【引数】 selected_line
-#   【戻り値】　なし
-#   【機能】 メンバ変数であるselected_lineを入力値に設定する．plot()をコールする．
-#
-#   selected(event)
-#   【引数】 event
-#   【戻り値】　なし
-#   【機能】　テーブルの選択イベントが発生した場合，本関数がコールされる．バインドはload_dile()で設定される．本関数ではset_selected_lineをコールする．
-#　　　　　　　　table.parent=1かつ x_table.parent=0の場合，tableで選択された行と同じ行を，x_tableに選択させる．(X-YテーブルとU-Vテーブルの連動選択時の処理である．）
-#
-#   set_offset_dist(float offset_dist)
-#   【引数】　offset_dist
-#   【戻り値】　なし
-#   【機能】　tableに表示されているライン全てのoffset_distを入力値に更新する．plot()，table_reload()をコールし，グラフ，表を更新する．
-#
-#   Change_CutDir()
-#   【引数】 なし
-#   【戻り値】　なし
-#   【機能】　tableで選択されているラインの向きをすべて逆転させる．plot()，table_reload()をコールし，グラフ，表を更新する．
-#
-#   Change_OffsetDir()
-#   【引数】 なし
-#   【戻り値】 なし
-#   【機能】 tableで選択されているラインのオフセット方向をすべて逆転させる．plot()，table_reload()をコールし，グラフ，表を更新する．
-#
-#   Swap_Selected_line()
-#   【引数】 なし 
-#   【戻り値】 list selected_items
-#   【機能】 tableで選択されている２本のラインのカット順を入れ替える．選択が2本である場合，Swap_lineをコールする．plot()，table_reload()をコールし，グラフ，表を更新する．
-#　　　　　　　　戻り値として，選択されたラインの番号をリストとして返す．（上位関数で入れ替えが成功したかを通知するために使用する．）
-#
-#   Swap_line(int item_num0, int item_num1):
-#   【引数】　int item_num0, item_num1
-#   【戻り値】　なし
-#   【機能】 入力された2本のラインについて，line_num_listの値を入れ替える．set_selected_lineをコールし，line_num1をselected_lineに設定する．
-#
-#   delete_Selected_line()
-#   【引数】 なし
-#   【戻り値】　なし
-#   【機能】　tableで選択されているラインをすべて非表示とする．delete_lineをコールする．plot()，table_reload()をコールし，グラフ，表を更新する．
-#
-#   delete_line(int item_num)
-#   【引数】　int item_num
-#   【戻り値】　なし
-#   【機能】　item_numに対応するラインのlin_num_listの値を-1に上書きする．tableからitem_numを削除する．
-#
-#   reverse_all()
-#   【引数】 なし
-#   【戻り値】　なし
-#   【機能】　テーブルで表示しているラインの順番を逆転させ，カット方向を逆転させる．
-#
-#   SortLine(float ox0, float oy0)
-#   【引数】 
-#   【戻り値】 なし
-#   【機能】 テーブルで表示しているラインを下記のルールに基づき並び替え，要すればカット方向，オフセット方向を変更する．
-#　　　　　　　　1. ox, oyから最も近い位置にあるライン端を検索する
-#　　　　　　　　2. 1.のライン端を有するラインを，カット番号0に設定する.1.のライン端がラインの終端である場合，カット方向，オフセット方向を逆転させる．
-#　　　　　　　　3. 1.のライン端の逆端から最も近い位置にあるライン端を検索する．        
-#　　　　　　　　4.　3.のライン端を有するラインを，カット番号1に設定する．3.のライン端がラインの終端である場合，カット方向，オフセット方向を逆転させる.
-#　　　　　　　　5. 3.のライン端から最も近い位置にあるライン端を検索する．
-#　　　　　　　　6. 3～5を，テーブルに表示されている全てのラインに対して実施する．
-#　　　　　　　　
-
-class dxf_file:
-    def __init__(self, ax, canvas, table, x_table, name):
-        self.ax = ax
-        self.canvas = canvas
-        self.table = table
-        self.x_table = x_table
-        self.name = name
-        self.line_list = []
-        self.line_num_list = []
-        self.selected_line = 0
-    
-    
-    def load_file(self, filename):
-        self.table.reset()
-        self.filename = filename
-        self.line_list = []
-        self.line_num_list = []
-        self.selected_line = 0
-        self.reload()
-        self.plot()
-        self.table.table.bind("<<TreeviewSelect>>", self.selected)
-        self.table.table.loaded_item_num = len(self.table.table.get_children())
-
-
-    def reload(self):
-        dwg = ez.readfile(self.filename)
-        modelspace = dwg.modelspace()
-        line_obj = modelspace.query('LINE')
-        spline_obj = modelspace.query('SPLINE')
-        arc_obj = modelspace.query('ARC') #ver2.2追加
-        poly_obj = modelspace.query('LWPOLYLINE') #ver2.2追加
-        
-        i = 0
-        while i < len(spline_obj):
-            temp_spline = spline_obj[i]
-            temp_spline_data = np.array(temp_spline.control_points)[:]
-            temp_line_object = line_object(temp_spline_data[:,0], temp_spline_data[:,1], i) 
-            self.line_list.append(temp_line_object)
-            self.line_num_list.append(i)
-            self.table.table.insert("", "end", values=(temp_line_object.num, temp_line_object.cut_dir, \
-                                                       temp_line_object.offset_dir, format(temp_line_object.offset_dist, '.2f'),\
-                                                       temp_line_object.line_type, int(temp_line_object.cutspeed_work)))
-            i += 1
- 
-        #ver2.2追加　円弧，ポリラインの読み込み　ここから
-        i_arc = 0        
-        while i_arc < len(arc_obj):
-            temp_arc = arc_obj[i_arc]
-            temp_arc_data = arc_to_spline(temp_arc)
-            temp_line_object = line_object(temp_arc_data[:,0], temp_arc_data[:,1], i + i_arc) 
-            self.line_list.append(temp_line_object)
-            self.line_num_list.append(i + i_arc)
-            self.table.table.insert("", "end", values=(temp_line_object.num, temp_line_object.cut_dir, \
-                                                       temp_line_object.offset_dir, format(temp_line_object.offset_dist, '.2f'), \
-                                                           temp_line_object.line_type, int(temp_line_object.cutspeed_work)))
-            i_arc += 1        
-        i = i + i_arc
-
-        i_poly = 0
-        while i_poly < len(poly_obj):
-            temp_poly = poly_obj[i_poly]
-            temp_poly_data = poly_to_spline(temp_poly)
-            temp_line_object = line_object(temp_poly_data[:,0], temp_poly_data[:,1], i + i_poly) 
-            temp_line_object.interp_mode = "linear" #poly_lineであることを設定する
-            self.line_list.append(temp_line_object)
-            self.line_num_list.append(i + i_poly)
-            self.table.table.insert("", "end", values=(temp_line_object.num, temp_line_object.cut_dir, \
-                                                       temp_line_object.offset_dir, format(temp_line_object.offset_dist, '.2f'), \
-                                                           temp_line_object.line_type, int(temp_line_object.cutspeed_work)))
-            i_poly += 1    
-        i = i + i_poly
-        #ver2.2追加　ここまで
-
-        j = 0
-        k = 0
-        while j < len(line_obj):
-            temp_line = line_obj[j]
-            temp_line_data = []
-            temp_line_data.append(temp_line.dxf.start)
-            temp_line_data.append(temp_line.dxf.end)
-            temp_line_data = np.array(temp_line_data)[:,0:2]
-            if norm(temp_line_data[0,0],temp_line_data[0,1],temp_line_data[1,0],temp_line_data[1,1]) != 0:
-                temp_line_object = line_object(temp_line_data[:,0], temp_line_data[:,1], i+k) 
-                self.line_list.append(temp_line_object)
-                self.line_num_list.append(i+k)
-                self.table.table.insert("", "end", values=(temp_line_object.num, temp_line_object.cut_dir, \
-                                                           temp_line_object.offset_dir, format(temp_line_object.offset_dist, '.2f'), \
-                                                               temp_line_object.line_type, int(temp_line_object.cutspeed_work)))
-                k += 1
-            j += 1
-            
-
-    def table_reload(self):
-        table_item_list = self.table.table.get_children()
-        i = 0
-        while i < len(table_item_list):
-            temp_item_num = table_item_list[i]
-            line_num = self.line_num_list[item2num(temp_item_num)]
-            temp_line_object = self.line_list[line_num]
-            self.table.table.item(temp_item_num, values=(temp_line_object.num, temp_line_object.cut_dir, \
-                                                         temp_line_object.offset_dir, format(temp_line_object.offset_dist, '.2f'), \
-                                                         temp_line_object.line_type, int(temp_line_object.cutspeed_work)))
-            i += 1
-
-    
-    def plot(self):
-        self.ax.clear()
-        self.ax.set_title(self.name)
-        self.ax.set_aspect('equal')
-        i = 0
-        while i < len(self.line_num_list):
-            line_num = self.line_num_list[i]
-            if line_num >= 0:
-                temp_line_object = self.line_list[line_num]
-                temp_line_object.update()
-                if temp_line_object.line_type == "line":
-                    col = "b"
-                else:
-                    col = "b"
-                    
-                if line_num == self.selected_line:
-                    self.ax.plot(temp_line_object.x, temp_line_object.y, color = col)
-                    X,Y = temp_line_object.x[0], temp_line_object.y[0]
-                    U,V = temp_line_object.x[1], temp_line_object.y[1]
-                    self.ax.quiver(X,Y,U-X,V-Y,color = col)
-                    if temp_line_object.cut_dir == "F":
-                        X,Y = temp_line_object.x_raw[0], temp_line_object.y_raw[0]
-                        U,V = temp_line_object.x[0], temp_line_object.y[0]
-                    else:
-                        X,Y = temp_line_object.x_raw[-1], temp_line_object.y_raw[-1]
-                        U,V = temp_line_object.x[0], temp_line_object.y[0]            
-                    if norm(X,Y,U,V) != 0:
-                        self.ax.quiver(X,Y,U-X,V-Y,color = "y") #ver2.2 バグ修正 Y を y に変更
-                else :
-                    self.ax.plot(temp_line_object.x, temp_line_object.y ,color = col, alpha = 0.2)    
-            i += 1
-        self.canvas.draw()
-        
-        
-    def set_selected_line(self, selected_line):
-        self.selected_line = selected_line
-        self.plot()
-    
-    
-    def selected(self, event):
-        selected_items = self.table.table.selection()
-        if not selected_items:
-            return None
-        item_num = item2num(selected_items[0])
-        item_index = self.table.table.get_children().index(selected_items[0])
-        line_num = self.line_num_list[item_num]
-        if self.table.parent == 1:
-            if self.x_table.parent == 0:            
-                try:
-                    x_item = self.x_table.table.get_children()[item_index]
-                    self.x_table.table.selection_set(x_item)
-                    self.x_table.table.see(x_item)
-                except:
-                    pass
-        self.set_selected_line(line_num)
-        
-        
-    def set_offset_dist(self, offset_dist):
-        table_item_list = self.table.table.get_children()
-        i = 0
-        while i < len(table_item_list):
-            temp_item_num = table_item_list[i]
-            line_num = self.line_num_list[item2num(temp_item_num)]
-            temp_line_object = self.line_list[line_num]
-            temp_line_object.offset_dist = offset_dist
-            temp_line_object.update()
-            i += 1   
-        self.table_reload()
-        self.plot()
-    
-    def Change_CutDir(self):
-        selected_items = self.table.table.selection()
-        i = 0
-        while i < len(selected_items):
-            temp_item_num = selected_items[i]
-            line_num = self.line_num_list[item2num(temp_item_num)]
-            temp_line_object = self.line_list[line_num]
-            temp_line_object.toggle_cut_dir()
-            temp_line_object.update()
-            self.table.table.item(temp_item_num, values=(temp_line_object.num, temp_line_object.cut_dir, \
-                                                         temp_line_object.offset_dir, format(temp_line_object.offset_dist, '.2f'), \
-                                                             temp_line_object.line_type, int(temp_line_object.cutspeed_work)))
-            i += 1   
-        self.table_reload()
-        self.plot()
-    
-    
-    def Change_OffsetDir(self):
-        selected_items = self.table.table.selection()
-        i = 0
-        while i < len(selected_items):
-            temp_item_num = selected_items[i]
-            line_num = self.line_num_list[item2num(temp_item_num)]
-            temp_line_object = self.line_list[line_num]
-            temp_line_object.toggle_offset_dir()
-            temp_line_object.update()
-            self.table.table.item(temp_item_num, values=(temp_line_object.num, temp_line_object.cut_dir, \
-                                                         temp_line_object.offset_dir, format(temp_line_object.offset_dist, '.2f'), \
-                                                             temp_line_object.line_type, int(temp_line_object.cutspeed_work)))
-            i += 1      
-        self.table_reload()
-        self.plot()
-
-
-    def Swap_Selected_line(self):
-        selected_items = self.table.table.selection()
-        if len(selected_items) == 2:
-            swap_line_num0 = item2num(selected_items[0])
-            swap_line_num1 = item2num(selected_items[1])
-            self.Swap_line(swap_line_num0, swap_line_num1)
-            self.table_reload()
-            self.plot()
-        return selected_items
-    
-
-    def Swap_line(self, item_num0, item_num1):
-        self.line_num_list[item_num0], self.line_num_list[item_num1] = self.line_num_list[item_num1], self.line_num_list[item_num0] 
-        self.set_selected_line(item_num1)    
-
-
-    def Merge_Selected_line(self):
-        selected_items = self.table.table.selection()
-        error = 0   # 0: 成功, 1: 選択数が誤り, 2: 隣接していない
-        if len(selected_items) == 2:
-            parent_line_num = item2num(selected_items[0])
-            child_line_num = item2num(selected_items[1])
-            if self.Merge_line(parent_line_num, child_line_num) == True:
-                self.delete_line(selected_items[1])
-                self.table_reload()
-                self.plot()
-                error = 0
-                return error, [selected_items[0]]
-            else:
-                error = 2
-                return error, selected_items
-            
-        elif len(selected_items) == 1:
-            error = 1
-            return error, [selected_items]
-        
-        else:
-            error = 1
-            return error, selected_items
-
-
-    def Merge_line(self, parent_item_num, child_item_num):
-        parent_line = self.line_list[parent_item_num]
-        child_line = self.line_list[child_item_num]
-        
-        x_p_st = parent_line.st[0]
-        y_p_st = parent_line.st[1]
-        x_c_st = child_line.st[0]
-        y_c_st = child_line.st[1]
-
-        x_p_ed = parent_line.ed[0]
-        y_p_ed = parent_line.ed[1]
-        x_c_ed = child_line.ed[0]
-        y_c_ed = child_line.ed[1]
-        
-        x_p = parent_line.x_dxf.tolist()
-        y_p = parent_line.y_dxf.tolist()
-        x_c = child_line.x_dxf.tolist()
-        y_c = child_line.y_dxf.tolist()     
-        
-        new_x = []
-        new_y = []
-
-        if norm(x_p_ed, y_p_ed, x_c_st, y_c_st) < LINE_MARGE_NORM_MN:
-            # 親ラインに子ラインを接続
-            new_x.extend(x_p)
-            new_y.extend(y_p)
-            new_x.extend(x_c)
-            new_y.extend(y_c)
-        elif norm(x_p_ed, y_p_ed, x_c_ed, y_c_ed) < LINE_MARGE_NORM_MN:
-            # 子ラインを反転させたのち、親ラインに子ラインを接続
-            x_c.reverse()
-            y_c.reverse()
-            new_x.extend(x_p)
-            new_y.extend(y_p)
-            new_x.extend(x_c)
-            new_y.extend(y_c)
-        elif norm(x_p_st, y_p_st, x_c_ed, y_c_ed) < LINE_MARGE_NORM_MN:
-            # 子ラインに親ラインを接続
-            new_x.extend(x_c)
-            new_y.extend(y_c)
-            new_x.extend(x_p)
-            new_y.extend(y_p)
-        elif norm(x_p_st, y_p_st, x_c_st, y_c_st) < LINE_MARGE_NORM_MN:
-            # 子ラインを反転させたのち、子ラインに親ラインに接続
-            x_c.reverse()
-            y_c.reverse()
-            new_x.extend(x_c)
-            new_y.extend(y_c)
-            new_x.extend(x_p)
-            new_y.extend(y_p)
-            
-        else:
-            # 端点が隣接していない線同士であるので、結合すべきではない。
-            return False
-        
-        parent_line.reset_point(new_x, new_y, parent_line.offset_ox, parent_line.offset_oy)
-        return True    
-    
-    
-    def delete_Selected_line(self):
-        selected_items = self.table.table.selection()
-        i = 0
-        while i < len(selected_items):
-            self.delete_line(selected_items[i])
-            i += 1             
-        self.table_reload()
-        self.plot()
-        
-        
-    def delete_line(self, item_num):
-        line_num = self.line_num_list[item2num(item_num)]
-        index = self.line_num_list.index(line_num)
-        if self.line_num_list[index] >= 0:            
-            self.line_num_list[index] = -1
-        self.table.table.delete(item_num)
-    
-    
-    def reverse_all(self):
-        alaivable_line_num_list = np.array(np.array(self.line_num_list.copy())[np.where(np.array(self.line_num_list.copy()) >= 0)])
-        new_line_num_list = alaivable_line_num_list[-1::-1]
-        
-        i = 0
-        j = 0
-        while i < len(self.line_num_list):
-            if self.line_num_list[i] >= 0:
-                temp_num = new_line_num_list[j]
-                temp_line = self.line_list[temp_num]
-                temp_line.toggle_cut_dir()
-                self.line_num_list[i] = temp_num                
-                j += 1
-                i += 1
-            else:
-                i += 1
-        
-        self.table_reload()
-        self.plot()        
-        
-    
-    def SortLine(self, ox0, oy0):
-        i = 0
-        norm_mn = np.inf
-        temp_cut_dir = 'F'
-        x0 = ox0
-        y0 = oy0
-        alaivable_line_num_list = sorted(np.array(self.line_num_list.copy())[np.where(np.array(self.line_num_list.copy()) >= 0)])
-        new_line_num_list = []
-
-        while i < len(alaivable_line_num_list) :
-            norm_mn = np.inf
-            j = 0
-            while j < len(alaivable_line_num_list):
-                num1 = alaivable_line_num_list[j]
-                if num1 in new_line_num_list:
-                    pass
-                
-                else:
-                    temp_line1 = self.line_list[num1]
-                    temp_norm_st = norm(x0, y0, temp_line1.st[0], temp_line1.st[1])
-                    temp_norm_ed = norm(x0, y0, temp_line1.ed[0], temp_line1.ed[1])
-                    if min(temp_norm_st, temp_norm_ed) < norm_mn:
-                        temp_line_num1 = num1
-                        norm_mn = min(temp_norm_st, temp_norm_ed)                      
-                        if temp_norm_st < temp_norm_ed:
-                            temp_cut_dir = 'F'
-                        else:
-                            temp_cut_dir = 'R'
-                j += 1
-                
-            temp_line1 = self.line_list[temp_line_num1]
-            if temp_cut_dir == 'F':
-                x0 = temp_line1.ed[0]
-                y0 = temp_line1.ed[1]
-                if temp_line1.cut_dir != temp_cut_dir:
-                    temp_line1.toggle_offset_dir()
-                    
-            if temp_cut_dir == 'R':
-                x0 = temp_line1.st[0]
-                y0 = temp_line1.st[1]
-                if temp_line1.cut_dir != temp_cut_dir:
-                    temp_line1.toggle_offset_dir()
-                        
-            temp_line1.set_cut_dir(temp_cut_dir)
-            new_line_num_list.append(temp_line_num1)
-            i += 1
-        
-        i = 0
-        j = 0
-        while i < len(self.line_num_list):
-            if self.line_num_list[i] >= 0:
-                self.line_num_list[i] = new_line_num_list[j]
-                j += 1
-                i += 1
-            else:
-                i += 1
-        
-        self.table_reload()
-        self.plot()        
-
-
-    def offset_origin(self, offset_ox, offset_oy):
-        table_item_list = self.table.table.get_children()
-        i = 0
-        while i < len(table_item_list):
-            temp_item_num = table_item_list[i]
-            line_num = self.line_num_list[item2num(temp_item_num)]
-            temp_line_object = self.line_list[line_num]
-            temp_line_object.reset_point(temp_line_object.x_dxf, temp_line_object.y_dxf, offset_ox, offset_oy)
-            temp_line_object.update()
-            i += 1   
-        self.table_reload()
-        self.plot()
-        
-    
-
-###############   dxf_fileクラス   ここまで　　    　#########################################################################################
-#######################################################################################################################################     
-
-
-
-#######################################################################################################################################
-###############   messeage_windowクラス   ここから　　    　##################################################################################
-
-#【説明】
-#　tk.Textクラスに，set_messeageメソッドを追加したクラスである．
-#
-#【親クラス】
-#　tk.Text
-#
-#【メンバ変数】
-#   変数名          型                   説明
-#   root            tk.Frame            textウィンドウを配置するフレームのインスタンス
-#   width           int                 textウィンドウの幅（文字数で指定）
-#   height          int                 textウィンドウの高さ（文字列数で指定）
-#
-#【実装メソッド】
-#   __init__(tk.Frame root, int width, int height)
-#   【引数】 root, width, height
-#   【戻り値】　なし
-#   【機能】 初期化処理を実行する．メンバ変数width，heightを入力値に設定する．
-#
-#   set_messeage(str messeage):
-#   【引数】　str messeage
-#   【戻り値】　なし
-#   【機能】 textウインドウの最下列にmesseageを挿入する．最下列を表示する．
-#　　　　　　　　
-#
-
-class messeage_window(tk.Text):
-    def __init__(self, root, width, height):
-        self.width = width
-        self.height = height
-        super().__init__(root, width = self.width, height = self.height)
-    
-    def set_messeage(self, messeage):
-        self.insert("end", messeage)
-        self.see(self.index('end'))
-
-
-###############   messeage_windowクラス   ここまで　　    　##################################################################################
-#######################################################################################################################################  
 
 class config:
     def __init__(self):
@@ -1046,6 +97,7 @@ class config:
             self.MESSEAGE = "設定ファイルの読み込み成功\n"
             
         except:
+            traceback.print_exc()
             self.FILENAME_XY = "ファイル名を入力して下さい。"
             self.FILENAME_UV = "ファイル名を入力して下さい。"
             self.OX = 0.0
@@ -1077,428 +129,11 @@ class config:
             self.offset_function = generate_offset_function(x_data, y_data)
             self.MESSEAGE = "%sの読み込み成功\n"%file_path
         except:
+            traceback.print_exc()
             x_data = [1,1000]
             y_data = [0,0]
             self.offset_function = generate_offset_function(x_data, y_data)
             self.MESSEAGE = "溶け量ファイルの読み込み失敗\n"
-
-#======================================================================================================================================
-#            関数の実装
-#======================================================================================================================================
-
-
-
-#======================================================================================================================================
-#            汎用関数
-#======================================================================================================================================
-#
-#   offset_line(list x, list y, float d, str cut_dir, str interp_mode)
-#   【引数】　x, y, d, cut_dir
-#   【戻り値】　new_x, new_y
-#   【機能】 x,y の座標点列をdだけオフセットした点列を作成する．カット方向により，オフセット方向が変わらないように，cut_dirによってオフセット方向を変更する．
-#　　　　　　　　アルゴリズムは下記を実装する．(円弧補間は不要なので省略している)
-#          https://stackoverflow.com/questions/32772638/python-how-to-get-the-x-y-coordinates-of-a-offset-spline-from-a-x-y-list-of-poi
-#　　　　　　　　
-#   norm(float x0, float y0, float x, float y)
-#   【引数】 x0, y0, x, y
-#   【戻り値】　float 距離
-#   【機能】　(x0,y0) と(x,y)の距離を計算する．
-#
-#   norm_3d(float x0, float y0, float z0, float x, float y, float z)
-#   【引数】 x0, y0, z0, x, y, z
-#   【戻り値】 float 距離
-#   【機能】 (x0,y0,z0) と(x,y,z)の距離を計算する
-#
-#   item2num(str item_num)
-#   【引数】 item_num
-#   【戻り値】 int num
-#   【機能】 tk.treeのItem番号（I***，***がアイテム番号）から***を抽出し，int型に変換して出力する．***は16進法である．
-#
-#   generate_arc_length_points(line_object line_object, int N)
-#   【引数】 line_object, N
-#   【戻り値】　list x_p, y_p
-#   【機能】 line_objectのx, y点列をスプライン補完し，これをN等分した点列を作成する．
-#        
-#   generate_arc_length_points4line(float x_st, float  y_st, float x_ed, float y_ed, int N)
-#   【引数】　x_st, y_st, x_ed, y_ed, N
-#   【戻り値】　list x_p, y_p
-#   【機能】 (x_st, y_st), (x_ed, y_ed)を両端に持つ直線を，N等分した点列を作成する．
-#
-#   plot_3d_cut_path(Axes ax, list x, list y, list u, list v, float Z, int n_plot)
-#   【引数】 ax, x, y, u, v, Z, n_plot
-#   【戻り値】　list point_dist_list
-#   【機能】　axにx,y,u,vをプロットする．x,y平面とu,v平面の距離はZとする．各(x,y),(u,v)の点の距離を計算し，point_dist_list配列として出力する．
-#
-#   file_chk(str filename)
-#   【引数】 file_name
-#   【戻り値】 int 成功可否(1：読み取り成功, 0:拡張子が.dxfでない, -1:ファイル存在せず)
-#   【機能】 指定されたfilenameのファイルが存在するか，拡張子が.dxfかをチェックする．
-#
-#   gen_g_code_line_str(list x, list y, list u, list v)
-#   【引数】 x, y, u, v
-#   【戻り値】 list code_str
-#   【機能】 x, y, u, vの各座標からgコードを生成する．gコードはX,Y,Z,A軸を使用するとする，gコードは「G01 X** Y** U** V**」のフォーマットとする．（G01は移動指令）
-#
-#   arc_to_spline(ezdxf.entities.Arc arc_obj)
-#   【引数】 arc_obj
-#   【戻り値】 np.array [xp, yp]
-#   【機能】 円弧からスプラインの点列を作成する．点数は10degにつき1点とする．点数は最少で3点以上となる．
-#
-#   poly_to_spline(ezdxf.entities.LWPolyline poly_obj)
-#   【引数】 poly_obj
-#   【戻り値】 np.array [xp, yp]
-#   【機能】 ポリラインから点列を作成する．
-#
-
-def offset_line(x, y, d, cut_dir, interp_mode): #ver2.2 interp_mode 追加　ポリラインの場合，オフセット点を中点ではなく，点列を使用するため
-    new_x = []
-    new_y = []
-    k = 0
-    sign = 1
-    sign_cut_dir = 1
-    
-    if cut_dir == "R":
-        sign_cut_dir = -1
-    
-    if len(x) < 2: 
-        return x, y
-
-    if len(x) == 2:
-        if x[0] == x[1]:
-            k = np.pi/2.0
-        else:
-            k = np.arctan((y[1]-y[0])/(x[1]-x[0]))
-        
-        new_x.append(x[0] - d*np.sin(k))
-        new_x.append(x[1] - d*np.sin(k))
-        new_y.append(y[0] + d*np.cos(k))
-        new_y.append(y[1] + d*np.cos(k))
-
-    if len(x) > 2:
-        if interp_mode == "linear" and len(x)%2 == 0:
-            num = 0
-            while num < len(x)/2:
-                i = num * 2
-                if x[i] == x[i+1]:
-                    k = np.pi/2.0
-                else:
-                    k = np.arctan((y[i+1]-y[i])/(x[i+1]-x[i]))
-
-                new_x.append(x[i]   - d*np.sin(k))
-                new_x.append(x[i+1] - d*np.sin(k))
-                new_y.append(y[i]   + d*np.cos(k))
-                new_y.append(y[i+1] + d*np.cos(k))
-
-                num += 1
-            
-        else:
-            i = 0
-            while i < len(x):
-                if i == 0:
-                    k = np.arctan2((y[1]-y[0]), (x[1]-x[0]))
-
-                elif i == len(x) - 1:
-                    k = np.arctan2((y[i]-y[i-1]), (x[i]-x[i-1]))
-                    
-                else:
-                    k = np.arctan2((y[i+1]-y[i-1]), (x[i+1]-x[i-1]))
-
-                new_x.append(x[i] - d*np.sin(k) * sign_cut_dir)      
-                new_y.append(y[i] + d*np.cos(k) * sign_cut_dir)
-                i += 1
-    
-    new_x = np.array(new_x)
-    new_y = np.array(new_y)
-    
-    return new_x, new_y
-
-
-def norm(x0, y0, x, y):
-    return np.sqrt((x-x0)**2 + (y-y0)**2)
-
-
-def norm_3d(x0, y0, z0, x, y, z):
-    return np.sqrt((x-x0)**2 + (y-y0)**2 + (z-z0)**2)
-
-
-def item2num(item_num):
-    temp = item_num.replace("I","")
-    temp = '0x0%s'%temp
-    num = int(temp,0) - 1    
-    return num
-
-
-def generate_arc_length_points(line_object, N):
-    
-    N = int(N)
-    if N < 2:
-        N = 2
-
-    length_array = line_object.calc_length_array()
-    sum_length = length_array[-1]
-    
-    t_p = length_array/sum_length
-    
-    x = line_object.x
-    y = line_object.y
-    
-    if line_object.line_type == "point":  
-        x_p = x
-        y_p = y
-        
-    if line_object.line_type == "line":  
-        fx_t = intp.interp1d(t_p, x, kind = "linear")
-        fy_t = intp.interp1d(t_p, y, kind = "linear")
-            
-        t_p_arc = np.linspace(t_p[0], t_p[-1], N)
-        
-        x_p = fx_t(t_p_arc)
-        y_p = fy_t(t_p_arc)
-
-    if line_object.line_type == "spline":
-        if line_object.interp_mode == "linear":
-            fx_t = intp.interp1d(t_p, x, kind = "linear")
-            fy_t = intp.interp1d(t_p, y, kind = "linear")
-                
-            t_p_arc = np.linspace(t_p[0], t_p[-1], N)
-            
-            t_p_arc_add_orgine_point = np.append(t_p, t_p_arc)
-            t_p_arc_add_orgine_point = np.sort(t_p_arc_add_orgine_point)              
-                    
-            x_p = fx_t(t_p_arc_add_orgine_point)
-            y_p = fy_t(t_p_arc_add_orgine_point)
-            
-
-        else:
-            fx_t = intp.CubicSpline(t_p, x)
-            fy_t = intp.CubicSpline(t_p, y)
-            
-            t_p_arc = np.linspace(t_p[0], t_p[-1], N)
-            
-            x_p = fx_t(t_p_arc)
-            y_p = fy_t(t_p_arc)
-    
-    return x_p, y_p
-    
-
-def generate_arc_length_points4line(x_st, y_st, x_ed, y_ed, N):
-    
-    t_p = np.array([0,1])
-    fx_t = intp.interp1d(t_p, [x_st, x_ed], kind = "linear")
-    fy_t = intp.interp1d(t_p, [y_st, y_ed], kind = "linear")
-        
-    t_p_arc = np.linspace(t_p[0], t_p[-1], N)
-    
-    x_p = fx_t(t_p_arc)
-    y_p = fy_t(t_p_arc)
-    
-    return x_p, y_p
-
-
-# Ver2.1変更 　引数追加
-def plot_3d_cut_path(ax, x, y, u, v, xm, ym, um, vm, z_xy, z_uv, z_m, n_plot):
-    
-    point_dist_list = []
-    
-    try:
-        if len(x) == len(y) == len(u) == len(v) == len(xm) == len(ym) == len(um) == len(vm):
-            zm = np.ones(len(x)) * 0.0
-            z = np.ones(len(x)) * z_xy
-            w = np.ones(len(x)) * z_m - z_uv
-            wm = np.ones(len(x)) * z_m
-            
-            ax.plot(x, y, z, 'k')
-            ax.plot(u, v, w, 'k')
-            
-            ax.plot(xm, ym, zm, 'k')
-            ax.plot(um, vm, wm, 'k')
-            
-            
-            if n_plot < 1:
-                n_plot = 1
-            
-            num_per_plot = int(len(x)/n_plot)
-            if num_per_plot < 1:
-                num_per_plot = 1
-                
-            i = 0
-            j = 0
-            while i < len(x):
-                if j > num_per_plot:
-                    ax.plot([x[i],u[i]], [y[i],v[i]],  [z_xy, z_m - z_uv], color = 'r', alpha = 0.5)
-                    ax.plot([x[i],xm[i]],[y[i],ym[i]], [z_xy, 0] , color = 'b', alpha = 0.5)
-                    ax.plot([u[i],um[i]],[v[i],vm[i]], [z_m - z_uv, z_m], color = 'b', alpha = 0.5)
-                    
-                    j = 0
-                temp_norm = norm_3d(xm[i], ym[i], 0, um[i], vm[i], z_m)
-                point_dist_list.append(temp_norm)
-                j += 1
-                i += 1
-            
-            return np.array(point_dist_list)                  
-            
-        
-        else:
-            return np.array([])
-    except:
-        return np.array([])
-        pass
-
-
-def file_chk(filename):
-    if op.exists(filename) == True:
-        if filename.split('.')[-1] == "dxf":
-            return 1
-        else:
-            return 0
-    else:
-        return -1
-
-
-#Ver2.0　変更 Gコード出力形式
-def gen_g_code_line_str(x,y,u,v, cut_speed):
-    code_str = ""
-    if len(x) == len(y) == len(u) == len(v):
-        i = 0
-        while i < len(x):
-            code_str += "G01 X%s Y%s U%s V%s F%s\n"%(format(x[i], '.6f'), format(y[i], '.6f'), \
-                                                     format(u[i], '.6f'), format(v[i], '.6f'), \
-                                                         format(cut_speed, '.1f'))
-            i += 1
-        return code_str
-
-
-#ver2.2 追加　円弧をスプライン化する機能の追加
-def arc_to_spline(arc_obj): #ezdxf arcオブジェクト
-    
-    radius = arc_obj.dxf.radius
-    x_center_point  = arc_obj.dxf.center[0]
-    y_center_point  = arc_obj.dxf.center[1]
-    start_angle     = arc_obj.dxf.start_angle
-    end_angle       = arc_obj.dxf.end_angle
-    
-    if end_angle < start_angle:
-        end_angle += 360
-        
-    num_point = int(np.abs(end_angle - start_angle)/10.0)
-    if num_point < 3:
-        num_point = 3
-    angle_array = np.radians(np.linspace(start_angle, end_angle, num_point))
-    x_p = radius * np.cos(angle_array) + x_center_point
-    y_p = radius * np.sin(angle_array) + y_center_point
-    
-    return np.array([x_p, y_p]).T
-
-
-#ver2.2　追加 ポリラインを点列化する機能の追加 
-def poly_to_spline(poly_obj):
-    point_x = np.array(poly_obj.get_points())[:,0]
-    point_y = np.array(poly_obj.get_points())[:,1]
-    new_point_x = [point_x[0]]
-    new_point_y = [point_y[0]]
-    
-    i = 1
-    while i < len(point_x)-1:
-        new_point_x.append(point_x[i])
-        new_point_x.append(point_x[i])
-        
-        new_point_y.append(point_y[i])
-        new_point_y.append(point_y[i])
-        
-        i += 1
-
-    new_point_x.append(point_x[-1])
-    new_point_y.append(point_y[-1])
-    
-    return np.array([new_point_x, new_point_y]).T
-
-
-def get_curdir():
-    # PyInstallerで実行されているかどうかをチェック
-    if getattr(sys, "frozen", False):
-        # EXEの実行ファイルのパスを取得
-        curdir = os.path.dirname(sys.executable)
-    else:
-        # スクリプトの実行ファイルのパスを取得
-        curdir =  os.path.dirname(os.path.abspath(__file__))
-    return curdir
-
-
-def generate_offset_function(x_array, y_array):
-    if x_array[0] > 0:
-        x_array_func = np.append(0, x_array)
-        y_array_func = np.append(y_array[0], y_array)
-    
-    if x_array[-1] < 10000:
-        x_array_func = np.append(x_array, 10000)
-        y_array_func = np.append(y_array, y_array[-1])   
-    
-    offset_function = intp.interp1d(x_array_func, y_array_func, kind = "linear")
-    
-    return offset_function
-
-
-def get_cross_point(p1_x, p1_y, p2_x, p2_y, p3_x, p3_y, p4_x, p4_y):
-    # https://imagingsolution.blog.fc2.com/blog-entry-137.html
-    s1 = ((p4_x - p2_x) * (p1_y - p2_y) - (p4_y - p2_y) * (p1_x - p2_x)) / 2.0
-    s2 = ((p4_x - p2_x) * (p2_y - p3_y) - (p4_y - p2_y) * (p2_x - p3_x)) / 2.0
-    
-    c1_x = p1_x + (p3_x - p1_x) * (s1 / (s1 + s2))
-    c1_y = p1_y + (p3_y - p1_y) * (s1 / (s1 + s2))
-    
-    return c1_x, c1_y
-
-
-def generate_offset_interporate_point(l0_x, l0_y, l1_x, l1_y, l0_offset, l1_offset):
-    # https://w3e.kanazawa-it.ac.jp/math/category/vector/henkan-tex.cgi?target=/math/category/vector/naiseki-wo-fukumu-kihonsiki.html&pcview=2
-    try:
-        a1 = l0_x[1] - l0_x[0]
-        a2 = l0_y[1] - l0_y[0]
-        b1 = l1_x[0] - l1_x[1]
-        b2 = l1_y[0] - l1_y[1]
-            
-        sita = np.arccos( (a1*b1 + a2*b2)/(np.sqrt(a1**2 + a2**2) * np.sqrt(b1**2 + b2**2)) ) / 2.0
-        beta = np.arctan2(a2, a1)
-        
-        R = min(l0_offset, l1_offset)
-        
-        cx, cy = get_cross_point(l0_x[0], l0_y[0], l1_x[0], l1_y[0], l0_x[1], l0_y[1], l1_x[1], l1_y[1])
-        
-        p1_x = cx - R * (1/np.tan(sita)) * np.cos(-beta) 
-        p1_y = cy + R * (1/np.tan(sita)) * np.sin(-beta) 
-        
-        p0_x = p1_x - R*np.sin(-beta)
-        p0_y = p1_y - R*np.cos(-beta)
-    
-        sita_array = np.linspace(np.pi/2 + beta, 2*sita + beta - np.pi/2, N_FILLET_INTERPORATE)
-    
-        x_intp = R*np.cos(sita_array) + p0_x
-        y_intp = R*np.sin(sita_array) + p0_y
-    except:
-        traceback.print_exc()
-        #内積・外積が計算できない場合は、補完しない直線を返す
-        x_intp = np.array([l0_x[1], l1_x[0]])
-        y_intp = np.array([l0_y[1], l1_y[0]])     
-        pass
-        
-    
-    #for debug
-    """
-    print(np.degrees(sita), np.degrees(beta))
-    plt.plot(cx, cy, "ro")
-    plt.plot(p1_x, p1_y, "bo")
-    plt.plot(p0_x, p0_y, "go")
-    plt.plot(l0_x, l0_y, "-o")
-    plt.plot(l1_x, l1_y, "-o")
-    plt.plot(x_intp, y_intp)
-    ax = plt.gca()
-    ax.set_aspect('equal', adjustable='box')
-    """
-    
-    return x_intp, y_intp
-    
-
-
 
 #======================================================================================================================================
 #            ボタンにより呼び出される関数
@@ -1686,6 +321,7 @@ def Set_OffsetDist(dxf_obj0, dxf_obj1, entry, messeage_window):
         dxf_obj1.set_offset_dist(OffsetDist)
         messeage_window.set_messeage("オフセット距離を%sに設定しました。\n"%OffsetDist)
     except:
+        traceback.print_exc()
         messeage_window.set_messeage("実数値を入力して下さい。\n")
         pass
 
@@ -1716,6 +352,7 @@ def AutoLineSort(dxf_obj0, dxf_obj1, entry_x, entry_y, messeage_window):
         dxf_obj1.SortLine(ox, oy)
         messeage_window.set_messeage("自動整列しました。\n")
     except:
+        traceback.print_exc()
         messeage_window.set_messeage("実数値を入力して下さい。\n")
         pass
 
@@ -1877,6 +514,7 @@ def Set_OffsetDistFromFunction(dxf_obj0, dxf_obj1, entry_XYDist, entry_UVDist, e
             messeage_window.set_messeage("XY座標とUV座標でライン数が一致しません。XY：%s本，UV：%s本\n"%(len(a_line_num_list0),len(a_line_num_list1)))
 
     except:
+        traceback.print_exc()
         messeage_window.set_messeage("入力値に誤りがあります。オフセット値更新を中止しました。\n")
 
 # Ver2.1変更　引数追加，距離別指定可能
@@ -2031,6 +669,7 @@ def gen_g_code(dxf_obj0, dxf_obj1, entry_ox, entry_oy, entry_ex, entry_ey, entry
         if temp_error_flg == True:
             messeage_window.set_messeage("入力値に誤りがあります。Gコード生成を中止しました。\n\n")
     except:
+        traceback.print_exc()
         messeage_window.set_messeage("Gコード生成途中でエラーが発生しました。\n\n")
         pass
 
@@ -2240,6 +879,7 @@ def offset_origin(dxf_obj0, dxf_obj1, entry_offset_ox, entry_offset_oy, messeage
         dxf_obj1.offset_origin(offset_ox, offset_oy)
         messeage_window.set_messeage("XY,UV座標をX:%s, Y:%sオフセットしました。\n"%(offset_ox, offset_oy))
     except:
+        traceback.print_exc()
         messeage_window.set_messeage("オフセット中にエラーが発生しました\n")
