+++ src/cam_global.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Mon Aug 11 15:53:27 2025
+
+@author: hirar
+"""
+
+#======================================================================================================================================
+#            グローバル変数
+#======================================================================================================================================
+VERSION = "3.0"           #バージョン
+Z_ERROR = 5.0             #2020.10.13　ver2.1 新規追加  単位：mm Z面の許容誤差 
+LINE_MARGE_NORM_MN = 1    #単位:mm ラインを結合時にラインを結合してよいかを判断するためのライン端点間距離
+N_FILLET_INTERPORATE = 10 #オフセットした線間を補完するフィレットの座標点数
